requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
dataclasses-json==0.5.7
Deprecated==1.2.13
Django==3.2.14
djangorestframework==3.13.1
Fiona==1.8.21
geopandas==0.10.2
idna==3.3
importlib-metadata==4.12.0
kombu==5.2.4
marshmallow==3.17.0
marshmallow-enum==1.5.1
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
prompt-toolkit==3.0.30
pyparsing==3.0.9
pyproj==3.2.1
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
Shapely==1.8.2
six==1.16.0
sqlparse==0.4.2
typing-inspect==0.7.1
typing_extensions==4.3.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
